+++ src/views/nodes/remotesNode.ts
@@ -28,7 +28,11 @@ export class RemotesNode extends ViewNode<RepositoriesView> {
             return [new MessageNode(this.view, this, 'No remotes could be found')];
         }
 
-        remotes.sort((a, b) => a.name.localeCompare(b.name, undefined, { numeric: true, sensitivity: 'base' }));
+        remotes.sort(
+            (a, b) =>
+                (a.default ? -1 : 1) - (b.default ? -1 : 1) ||
+                a.name.localeCompare(b.name, undefined, { numeric: true, sensitivity: 'base' })
+        );
         return [...Iterables.map(remotes, r => new RemoteNode(this.uri, this.view, this, r, this.repo))];
     }